KIALI-840 Hide sidebar on small windows
This is the naivest implementation possible, and might be a good idea to
review or find some other implementation that's smarter.

This aims to fix the following problem: when you resize the window to a
smaller size, the sidebar goes over the graph, so we hide it and
add it back when the windows is bigger.

Before:

![](https://i.imgur.com/SnDIJGJ.png)

After:

![](https://i.imgur.com/aEZgS7u.png)

There are some caveats that I could find:

* Graph dimensions are not responsive, so even removing the sidebar does
not fixes the size of the graph to fill the whole available screen
space. That causes some weird bugs, such as this one:

![](https://i.imgur.com/uJywPnS.png)

Ideas on how to fix this are welcome.