KIALI-1425 Add control to change reporter in the metrics detail pages
Showing metrics only for one reporter leads to some gaps in the available data. In order to fill the gaps, a dropdown is added to the metrics page to let the user choose the reporting entity to view the available metrics.

Related to KIALI-1421 KIALI-1422 which are about an issue with an inconsistency in the metrics being shown. This is not really an inconsistency, but a difference in the metrics, depending on the point of view of the reporter. The reporter could be the workload generating the traffic (source) or the workload receiving the traffic (destination).