KIALI-1525 Loading message when graph loads the first time
Loading message was already coded, but it wasn't being shown because CytoscapeGraph.shouldComponentUpdate wasn't allowing it to render when needed. A conditional is aded to let it render.

Also, graph state is now reset to initial values when graph namespace is changed. This allows the loading message to be shown, avoiding the graph of the old namespace to remain.