KIALI-698 Remove prettier from yarn start, yarn test and yarn build
Prettier fixes code syntax in order to comply with _good_ code stylings.
Good syntax is only relevant for developers/humans. So, it's not really
needed when running `yarn start`, `yarn test` nor `yarn build`, which
are meant to build the code (good syntax is irrelevant for
compilers/transpilers).

In CI, prettier is already called directly to trigger a failure if PRs
or code being merged/pushed. Removing prettier from build and test
stages will have the benefit that prettier will be called once in CI.

In dev envs, prettier is part of `yarn precommit` which can be
configured to interrupt a commit if prettier finds and fixes syntax
issues, allowing early detection before pushing to the repo.