KIALI-2894 Scroll to top of side panel when graph selection changes
https://issues.jboss.org/browse/KIALI-2894

This ensures that the side panel scroll is at the top position when changing a selection in the graph.

The issue only happens when changing selection of same cytoscape "type" (i.e select and edge while another edge is currently selected). It does not happen if screen resolution leads to no scrollbars in side panel when changing context.