Update to Typescript 3.5
Updates to latest version of typescript:
- https://devblogs.microsoft.com/typescript/announcing-typescript-3-5/

This update makes the compiler a little more strict, so some of the
things that we had before will not work. This update has the same
motivation as #1154, and having a more up to date typescript helps us
with it (some errors happen because of the lacking `unknown` type, which
is defined as `null | undefined | T extends {}`).

Some other changes:

* `@types/react-redux` has been updated because it generates a
  compilation error. This is build-only, so no impact. Also, we're already
  running the latest redux, so we're just keeping them in sync.
* `hoist-non-react-statics` had improper typescript typings and had to
  be forcefully updated. I did some investigations on the changes for
  the versions, nothing seems non-breaking.
* Health-related API namespaces were erasing types, so typescript just
  defaulted to `any` and ignored the typing. This has been fixed.
* Typescript 3.5 broke webpack ts-loader implementation when using
  `incremental`: https://github.com/microsoft/TypeScript/issues/31447
  This is fixed by also setting the `tsbuildinfo` file on tsconfig.

Some tests are failing, and I found the same problems on #1154, so I'm
backporting those changes to the tests.