KIALI-2965 Fix conflicting Action keys
Two actions had the same key: SET_INFO for grafana and for mtls, which resulted in a conflict ; grafana info being set with mtls data, causing the grafana links to be broken.

In order to not reproduce the same issue in the future, I'm putting all keys within the same enum.