KIALI-2529 Add type hints to composite types
Typescript has no way to know if the item has the health attribute or
not. Before, it just defaulted to any, but when we start using babel, it
becomes a compilation error.

This is one of those cases where the current implementation is incorrect
and only worked because it downcasted it to `any`.