Implement token authentication
This is providing a new login screen with only one "Token" field. The user is expected to paste a service account token to login into Kiali.

The UI will pass the provided token to the back-end. In turn, the back-end will provide a cookie as usual (if the Token is valid). The back-end should use the provided token during the session of the user to make calls to the Kubernetes/OpenShift API, so that he cluster's RBAC takes effect.

Fixes kiali/kiali#1139