KIALI-3150 Fix code splitting issue
When introducing code-splitting in Kiali, we are hitting an external bug or an unsupported scenario. The issue is either in PatternFly3 or in WebPack/Babel (or in something else).

For the NotificationDrawer component, PatternFly 3 uses the syntax that can be seen in: https://github.com/patternfly/patternfly-react/blob/patternfly-react%402.31.1/packages/patternfly-3/patternfly-react/src/components/Notification/NotificationDrawer/index.js#L21-L33

The PatternFly code is importing components and assigning them to a property of NotificationDrawer. This breaks the code-splitting mechanism and all those components assigned as properties of NotificationDrawer become undefined.

The issue is fixed by direct importing of those components in Kiali.