Implement mini-graph on detail pages
Start implementing mini-graph on detail pages.

Mini-graph for Applications is almost ready. There is a small issue when rendering the graph. Probably related to browser/components resizing.

The mini-graph for services and workloads is in draft state. Since the code is very similar, the mini-graph for apps will be finished first. Then, the implementation will be replicated to the other details pages.

Related to kiali/kiali#1960