Create MiniGraphCard component to reduce code duplication
The mini-graph is present in app, workload and service detail pages. The mini-graph related code is very similar on all these pages. This is extracting the mini-graph to a component to reduce code duplication.

Also, in CytoscapeGraph component, the `activeNamespaces` and `graphType` properties were removed because these data is present in the datasource.