Re-use mini-graph datasource in the traffic tab
The traffic tab is mostly using the same data as the mini-graph. Traffic tab is built from the graph's back-end endpoint. To save some trips to the network, these changes are re-using the data already loaded by the mini-graph datasource to build the list on the traffic tab.

This has two benefits:
* Less network usage, already mentioned
* Traffic tab is in sync with mini-graph

Related kiali/kiali#1960