Move the checkCredentials to AuthenticationController.tsx
The call to checkCredentials ThunkAction was being called unconditionally in the LoginPage to initiate the login cycle, if needed. Moving this call to the AuthenticationController.tsx for more centralized control of the login cycle, to call it only when needed instead to always call on load, and also to "fix" the login form being erroneously displayed shortly while the OAuth/OpenID cycle is being finalized.