Add navigation to remote Kiali in context menu
This fixes graph context menus for nodes that represent resources in non-local clusters. The context menu was trying to open a URL in the local Kiali, but the right action is to open a "remote Kiali instance" (if any) where the resource can be inspected. A dialog was added to warn the user that another Kiali instance is going to be accessed.

There are still a few issues:
* Current code assumes that the remote Kiali is always accessible. Some checks must be done to ensure that we can build a link to the remote Kiali. If a link cannot be built, the context menu should be disabled.
* Some option changes cause the cytoscape control to umount-remount. When this happens, the event handler that shows the modal dialog becomes disconnected, leaving the context menu unusable.