Drop context menu of non-local node if not accessible
When building the graph node context menu, options are made available only if the node belongs to the "local" cluster, or if the node belongs to a remote cluster where a Kiali was successfully discovered and is accessible. Otherwise, no options in the context menu are made available.

The clusters list had to be moved up to the server config that is fetched at UI boot time.