Fix issue when clicking a node
NodeContextMenu.tsx exports the getOptions function, which is reused at summary panel controls. It was assumed that it's only called if links are going to build properly, but that's not the case So, it now returns only link that can be built.

Additionally, links in side panel are now redirecting to the remote Kiali, if needed. Also, if remote Kiali is unreachable, a plain text is shown.