Show hosts for Ingress and VSs in graph as expandable
The graph will, initially, show the text "X hosts" on nodes that represent Ingress Gateways or nodes that have attached a VirtualService. Hovering the mouse's cursor over this text will expand the text to show what are those hosts limited to showing up to 5 hosts. If the node has associated more than 5 hosts, a legend stating this fact will appear at the end of the list.