Change the behavior of graph hide to be less aggressive about hiding nodes that aren't explicitly hit by the hide expression.  Previous strategy:
1) hide the nodes and edges explicitly hit by the hide expression
2) hide the still-visible edges connected to hidden nodes
3) hide the still visible-nodes that no longer have visible edges

This approach left no orphans by design, because orphans tend to stack
in the layout, and in most cases it seemed like hiding them helped the overall
behavior of graph-hide, providing cleaner results.

But, in certain cases, when the desired result is to see orphans, this
approach is too aggressive.  So, this commit changes the strategy to:

1) hide the nodes and edges explicitly hit by the hide expression
2) hide the still visible-nodes that no longer have visible edges
3) hide the still-visible edges connected to hidden nodes

This approach has the possibility of leaving many orphans, but it won't
hide a node that the hide expression really didn't hit.  With the new
layout doing a better job of handling orphan nodes, this may be better.